Move parse_maps into maps module
This change introduces a new internal module, maps, and moves the
existing /proc/.../maps parsing functionality into it.
